requirements.txt
robotframework
robotframework-appiumlibrary
robotframework-faker